change how we detect different types of files
ReadYaml uses yaml.Unmarshal which does not enforce that
the fields in the struct are present in the yaml file.
So even when unmarshalling random yaml files into a docker-compose
struct it doesn't give any errors. Errors are only produced in
rare circumstances where there is a type mismatch.

To fix this we add a new function ReadMove2KubeYaml that
is meant for use with our file formats. For file formats that we
don't control like docker-compose we defer to appropriate libraries.

Signed-off-by: Harikrishnan Balagopal <harikrishmenon@gmail.com>