feat: New architecture for Move2Kube using chained and containerizable transformers (#464)
The new architecture makes transformers as building blocks which can be chained to fit user requirements. Each transformer is responsible for a particular job, and then the other transformers carry on.

For example in compose, the first transformer detects the docker-compose files, the second stage fills the IR, and then kubernetes, tekton and other transformers create the final directory structure.

It also uses isolated environments for each transformers, so that they don't end up editing the source data. It also uses copy in and copy out of data in cases of containers using the environment concept.

The yaml transformation and parameterisation features will be added again in later PRs.

The Test cases are not yet rewritten for the new architecture. Hence build will fail till the testcases are revised.

Signed-off-by: Ashok Pon Kumar <ashokponkumar@gmail.com>