feat: add the parameterize command
Parameterizer allows parameterizing fields in the k8s resources.
Added a new module to parameterize k8s resource yamls.
The modules is also exposed as a command.

The input parameterizers and packaging yamls are provided in the config folder.
The format for the parameterizers is available in testdata folder of the
parameterizer module.

Also ignore the ruby dependency directories.

Signed-off-by: Harikrishnan Balagopal <harikrishmenon@gmail.com>