fix: bug where 2 questions about port are asked
When "intercepts" was changed to "passthrough" concept,
some of the built-in transformer yamls were not updated properly with the new field.
The WAR router transformer was running in parallel with the
WAR processor transformers (JBoss, Liberty, Tomcat).
This causes extra questions to be asked and ports 8080 and 9080 to come up together.

Signed-off-by: Harikrishnan Balagopal <harikrishmenon@gmail.com>