feat: add an ArgoCD transformer to generate the ArgoCD specific deployment yamls (#803)
* feat: add an ArgoCD transformer to generate the ArgoCD specific deployment yamls

Signed-off-by: Harikrishnan Balagopal <harikrishmenon@gmail.com>

* fix: update dependencies to fix build errors after adding argocd

Remove the useless -extldflags "-static" flag because that only applies
to the external linker `ld`. Golang uses the internal Go compiler and linker
which by default links everything statically.

The external compiler `clang` and external linker `ld` only gets invoked
when there is C code to be compiled. ArgoCD has some C code that needs to be
compiled.

Unfortunately there is no way to compile statically against system libraries
on most systems (except Linux) so the -extldflags "-static" flag causes a build
failure. So we need to remove that flag.

Another alternative to removing the -extldflags "-static" flag is to disable CGO
CGO_ENBALED=0 go build ..... does work but disables CGO for all packages. This
will only work if all packages provide Golang replacements for the C code that they
were using.

This could cause panics/issues down the line if the Golang replacements provided by
the various packages we use are mere stubs. It also will cause compilation failure
if such replacements are not available for a particular package.

There are more fine-grained ways to disable CGO for a single package like
https://pkg.go.dev/os/user for example.

"When cgo is available, and the required routines are implemented in libc
for a particular platform, cgo-based (libc-backed) code is used. This can be overridden
by using osusergo build tag, which enforces the pure Go implementation."

See:
https://www.reddit.com/r/golang/comments/53nl4b/why_all_go_binaries_are_staticallylinked/
https://lists.sr.ht/~sircmpwn/public-inbox/%3C3A12328E-5EC6-4E34-A676-64A3039A2AF7%40herrbischoff.com%3E
https://stackoverflow.com/questions/3801011/ld-library-not-found-for-lcrt0-o-on-osx-10-6-with-gcc-clang-static-flag
https://stackoverflow.com/questions/844819/how-to-static-link-on-os-x

Signed-off-by: Harikrishnan Balagopal <harikrishmenon@gmail.com>